add functions writing in files and some fixes

+++ template-engine/stringOperations.h
@@ -7,8 +7,8 @@ std::vector<std::string> split(const std::string&, char);
 std::vector<std::string> split(const std::string&, const std::string&);
 void replace(std::string&, const std::string&, size_t, size_t);
 long long findThreeSymbols(const std::string&, char first, char second, char third, size_t& i);
-size_t strToNum(const std::string& str);
-std::string numToStr(size_t num);
-size_t calcNumSize(size_t num);
+long long strToNum(const std::string& str);
+std::string numToStr(long long num);
+size_t calcNumSize(long long num);
 
 #endif //TEMPLATE_ENGINE_STRINGOPERATIONS_H
